### 2018-11-20 f81b022
Pipeline str needs to be passed instead of individual.
