+++ gama/genetic_programming/algorithms/async_ea.py
@@ -62,7 +62,8 @@ def async_ea(start_population, toolbox, evaluation_callback=None, restart_callba
                 if n_jobs > 1:
                     logger.flush_to_log(log)
                 log_parseable_event(log, TOKENS.EVALUATION_RESULT, individual.fitness.start_time,
-                                    individual.fitness.time, individual.fitness.wvalues, individual._id, individual)
+                                    individual.fitness.time, individual.fitness.wvalues,
+                                    individual._id, individual.pipeline_str())
 
                 if evaluation_callback:
                     _safe_outside_call(partial(evaluation_callback, individual), exceed_timeout)
