### 2018-11-28 #1088 @rhiever
I'd love to, but I haven't had much extra time for reviewing lately. Maybe @mfeurer? @ledell or someone on the H2O dev team? @weixuanfu?
