### 2018-11-28 #1088 @weixuanfu
I think it is a very interesting tool and I believe @PGijsbers did very good jobs related to autoML, but I didn't have much time for reviewing neither.
