### 2018-11-29 #1088 @arokem
@fmohr : would you like to review this article for [The Journal of Open Source Software](http://joss.theoj.org/)?
