### 2018-11-29 dc54359
Rework metric in code.
