### 2018-11-30 #1088 @fmohr
Hey, thanks for asking. In general, I am open to review for JOSS. In this particular case, however, I'm afraid I can't help out; I'm a Java guy and not at all an expert for Python. I don't know to which degree this is a problem, but Pieter mentions this as one of two criterions, so probably speaking Python should be a criterion.
