### 2018-11-30 #1088 @PGijsbers
I have to admit I'm not intimately familiar with the level of code inspection that is done in the JOSS review process. I listed Python as required since the package is written entirely in Python, I figured it would make sense to look for people who are familiar so as to save time in the review process. That said, I would probably feel the same reviewing a package written in Java.
