### 2018-11-30 54d8e07
Terminals and Primitives returned as List instead of Generator.
