### 2018-11-30 b83fe15
Redudant cast now that lists are returned.
