### 2018-11-30 b480545
Fixed a bug in eliminate test.
