### 2018-11-30 b5ce358
Refactor use of scoring/metric.
