### 2018-12-02 #1088 @arokem
Yes. That all makes sense. Thanks for your candor, @fmohr.

Let's keep looking for a reviewer that speaks Python...
