### 2018-12-02 #1088 @arokem
Hey @krother, @nilesh-patil : would either or both of you be willing to review this article for JOSS?
