### 2018-12-03 683e89c
Update documentation for metric.
