### 2018-12-03 5f9e257
Update docs for scoring.
