### 2018-12-17 #1132 @whedon
[REVIEW]: GAMA: Genetic Automated Machine learning Assistant
**Submitting author:** @PGijsbers (<a href="http://orcid.org/0000-0001-7346-8075">Pieter Gijsbers</a>)
**Repository:** <a href="https://github.com/PGijsbers/GAMA" target ="_blank">https://github.com/PGijsbers/GAMA</a>
**Version:** v0.1.0
