### 2019-01-10 408b414
Revert to scikit-learn 0.19.1.
