+++ gama/utilities/auto_ensemble.py
@@ -244,7 +244,7 @@ class EnsembleClassifier(Ensemble):
         self._label_encoder = label_encoder
 
         # For metrics that only require class labels, we still want to apply one-hot-encoding to average predictions.
-        self._one_hot_encoder = OneHotEncoder(categories='auto').fit(self._y_true.reshape(-1, 1))
+        self._one_hot_encoder = OneHotEncoder().fit(self._y_true.reshape(-1, 1))
 
         if self._metric.requires_probabilities:
             self._y_score = self._one_hot_encoder.transform(self._y_true.reshape(-1, 1)).toarray()
