### 2019-01-10 feae836
Revert to scikit-learn 0.19.1.
