+++ setup.py
@@ -5,7 +5,7 @@ from setuptools import setup, find_packages
 requirements = [
     'numpy>=1.14.0',
     'scipy>=1.0.0',
-    'scikit-learn==0.20.0',
+    'scikit-learn==0.19.1',
     'stopit>=1.1.1',
     'liac-arff>=2.2.2',
     'category-encoders>=1.2.8'
