### 2019-01-15 b16e3a6
Force numpy 1.14.3 as required by d3m environment.
