### 2019-01-15 4925cea
Only enter ensemble building if it is desired.
