### 2019-01-15 #1132 @PGijsbers
That's odd, I can not recreate the issue locally (tried a clean install on a ubuntu docker image). Are you sure that GAMA is installed to the correct virtual environment (if you are using one)?
