### 2019-01-15 404709d
Method required for d3m.
