### 2019-01-15 #1132 @jsgalan
I have a few suggestions to be made to the authors:

1- add a few examples (not just the minimal example, but all five examples) to the Github repository. Also please check that the examples is wrongly edited, missing the last part of the statement

![screen shot 2019-01-15 at 11 31 09 am](https://user-images.githubusercontent.com/971958/51194798-cd117300-18b9-11e9-8785-8bb2eb3d3c06.png)

2- Provide a more detailed examples and explanations of the following sections: **Using ARFF files**, **Logging**, **Log Visualization** and **GAMA Search Space Configuration**.

3- The authors could add in the manuscript a few lines describing in detai

l logging and log visualization aspects, which can be interesting for a user/reader.

4- Documentation should be complete, in the comparison with other AutoML tools a paragraph is missing.
![screen shot 2019-01-15 at 11 28 41 am](https://user-images.githubusercontent.com/971958/51195449-1615f700-18bb-11e9-9621-4da47737fc56.png)
