### 2019-01-16 a96d74c
Examples now accessible as separate python files. Dirty trick to copy over files for doc building.
