+++ examples/regression_example.py
@@ -1,0 +1,16 @@
+from sklearn.datasets import load_boston
+from sklearn.model_selection import train_test_split
+from sklearn.metrics import mean_squared_error
+from gama import GamaRegressor
+
+if __name__ == '__main__':
+    X, y = load_boston(return_X_y=True)
+    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=0)
+
+    automl = GamaRegressor(max_total_time=180)
+    print("Starting `fit` which will take roughly 3 minutes.")
+    automl.fit(X_train, y_train)
+
+    predictions = automl.predict(X_test)
+
+    print('MSE:', mean_squared_error(y_test, predictions))
