+++ docs/source/user_guide/examples/using_arff_files.rst
@@ -10,14 +10,8 @@ The example script can be run by using e.g.
 `breast_cancer_test.arff <https://github.com/PGijsbers/gama/tree/master/gama/tests/data/breast_cancer_test.arff>`_.
 The target should always be specified as the last column.
 
-::
 
-    from gama import GamaClassifier
-
-    automl = GamaClassifier(max_total_time=180)
-    automl.fit(arff_file_path='your_train_data.arff')
-
-    label_predictions = automl.predict(arff_file_path='your_test_data.arff')
-    probability_predictions = automl.predict_proba(arff_file_path='your_test_data.arff')
+.. file below is copied in by conf.py
+.. literalinclude:: /user_guide/examples/arff_example.py
 
 The GamaRegressor also has ARFF support.
