### 2019-01-16 5bc129d
[skip ci] Additional text about ARFF files.
