### 2019-01-16 af89b2c
[skip ci] Additional text about ARFF files.
