### 2019-01-16 943dcc3
[skip ci] Individual pipeline crashes during optimization, and weight assignment in ensembling are now printed at DEBUG level.
