### 2019-01-16 60cfb25
[skip ci] Don't store analysis log for examples.
