+++ examples/arff_example.py
@@ -4,7 +4,7 @@ if __name__ == '__main__':
     print("Make sure you adjust the file path if not executed from the examples directory.")
     file_path = "../tests/data/breast_cancer_{}.arff"
 
-    automl = GamaClassifier(max_total_time=180)
+    automl = GamaClassifier(max_total_time=180, keep_analysis_log=False)
     print("Starting `fit` which will take roughly 3 minutes.")
     automl.fit(arff_file_path=file_path.format('train'))
 
