### 2019-01-16 0cb03bd
Timeout raised during callback is annoying, but not something that merits a warning. It still should be easily observable, so I lowered it to INFO.
