+++ _sources/user_guide/things_to_know.rst.txt
@@ -70,17 +70,23 @@ The following script manually sets up GAMA to print to stdout (and ignores the b
 Running the above script display the GAMA version used, and the hyperparameter values on initialization::
 
     Using GAMA version 0.1.0.
-    GamaClassifier(cache_dir=None,verbosity=None,n_jobs=1,max_eval_time=300,max_total_time=180,population_size=50,random_state=None,optimize_strategy=(1, -1),objectives=('neg_log_loss', 'size'))
+    GamaClassifier(cache_dir=None,verbosity=None,n_jobs=1,max_eval_time=300,max_total_time=180,population_size=50,random_state=None,scoring='neg_log_loss'))
 
 Actual values or hyperparameter names may vary depending on the version of GAMA you are using.
 Make sure to set *both* the log level of the log and the stream handler.
+
+An overview the log levels:
+ - **debug**: messages for developers and computers. This includes output required for log visualization (see next section). Expect a lot of output.
+ - **info**: general information about the optimization process.
+ - **warning**: serious errors that do not prohibit GAMA from running to completion (but results could be suboptimal).
+ - **error**: errors which prevent GAMA from running to completion.
 
 Log Visualization
 *****************
 
 When using the default hyperparameters, GAMA will produce a log file called `gama.log`.
 This log file is structured so that important events are easy to parse.
-For an example and easy way to visualize information from the log file, take a look at [this notebook](https://github.com/PGijsbers/gama/blob/master/notebooks/GAMA%20Log%20Parser.ipynb).
+For an example and easy way to visualize information from the log file, take a look at `this notebook <https://github.com/PGijsbers/gama/blob/master/notebooks/GAMA%20Log%20Parser.ipynb>`_.
 
 Events and Observers
 ********************
