### 2019-01-16 15ec077
[skip ci] Update to reflect new example uses breast_cancer dataset.
