### 2019-01-16 #1132 @PGijsbers
Hi @jsgalan!
I processed your feedback best I could, but I still have some questions:
 1. What more detailed explanation are you missing for the [log visualization](https://github.com/PGijsbers/gama/blob/master/notebooks/GAMA%20Log%20Parser.ipynb)?
 2. We mentioned the log visualization and type of questions you can currently answer with it in our paper submission. What in particular were you missing or expected to be changed?

I changed log levels of some statements so that anything of at least `logging.INFO` level will not end with a stream of warning statements. The reported crashes in the statements were harmless to the optimization process, but I concede that it is not a nice user experience :)
For all other aspects, I updated the docs.

Thanks again for your time.
