### 2019-01-16 #1132 @jsgalan
Hi all,

Sorry for being so vague on my descriptions. I will reformulate:

1- **add a few examples (not just the minimal example, but all five examples) to the Github repository**.

1.1-Also please check that the examples is wrongly edited, missing the last part of the statement <-- this was done! [check!]

2- What more detailed explanation are you missing for the log visualization? <- can this link or this information be found somewhere on the initial Github webpage

3- We mentioned the log visualization and type of questions you can currently answer with it in our paper submission. What in particular were you missing or expected to be changed?

The article currently reads:

> In addition to its general use AutoML functionality, GAMA aims to serve AutoML researches as well. During the optimization process, GAMA keeps an extensive log of
> progress made. Using this log, insight can be obtained on the behaviour of the population of pipelines. It can answer questions such as which mutation operator is most
> effective, how fitness changes over time, and how much time each algorithm takes

I think is important to include in those lines regarding the aspects such as:

a) the figures/plots it produces. (a.e Fitness over number of evaluations vs No. evaluations, Pipeline length over number of evaluations vs No. evaluations and more importantly how Fitness over number of evaluations by main learner vs No. evaluations )

b) all the methods/models that can be learned (a.e NB, Decision Tree,Boosting, Random Forest, KNN, SVC, LogReg). This is not stated clearly in the documentation.

Sorry for being stubborn but I think this should be main focus of the article and be very well described in the Github website, showing all the full capabilities of the software and all the models that can be tested using the implementation.

After revising I think some extra homework...

4- There were .csv files generated  that are not well described anywhere

![screen shot 2019-01-16 at 1 35 00 pm](https://user-images.githubusercontent.com/971958/51271463-a45eab80-1995-11e9-92e0-861ea0cc30b2.png)

4.1- There were folders empty folders created with the same name of the .csv files.
4.2- Can the user specify the folder/csv files names instead of automatically assigning the date of the experimental run?

Hope this helps.
