### 2019-01-16 #1132 @PGijsbers
> add a few examples (not just the minimal example, but all five examples) to the Github repository.

Could you elaborate what you mean specifically? I made three of them into separate scripts in the [example folder](https://github.com/PGijsbers/gama/tree/master/examples) and linked them from the `README.md`. What other examples are you referring to? [Setting log output](https://pgijsbers.github.io/gama/user_guide/index.html#logging) and adding an [event hook](https://pgijsbers.github.io/gama/user_guide/index.html#events-and-observers)?

> can this link or this information be found somewhere on the initial Github webpage

It is referred to in the documentation but I will add it to the `README.md`, good call.

> 4, 4.1

I turned it off for examples now, as I don't think examples should produce extra files. I will add documentation regarding the files should you wish to produce them.

>Can the user specify the folder/csv files names instead of automatically assigning the date of the experimental run?

Yes, using the 'cache_dir' hyperparameter when initializing a `GamaClassifier` or `GamaRegressor` object. I will make sure it also shows up in the regular documentation (not just API docs).

I'll make the changes tomorrow.
