### 2019-01-17 cde7a38
Specify master branch for images.
