+++ code_of_conduct.md
@@ -55,7 +55,7 @@ further defined and clarified by project maintainers.
 ## Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported by contacting the project team at ![my e-mail address](https://raw.githubusercontent.com/PGijsbers/gama/images/mail.png). 
+reported by contacting the project team at ![my e-mail address](https://raw.githubusercontent.com/PGijsbers/gama/master/images/mail.png). 
 All complaints will be reviewed and investigated and will result in a response that
 is deemed necessary and appropriate to the circumstances. The project team is
 obligated to maintain confidentiality with regard to the reporter of an incident.
