### 2019-01-17 6d260de
More info on used algorithms.
