### 2019-01-17 7d52480
Added visualization example to paper.
