### 2019-01-17 ecce143
[skip ci] Expanded docs with file info and use :code: as default interpreted text.
