### 2019-01-17 #1132 @PGijsbers
Hi @jsgalan,

> add a few examples (not just the minimal example, but all five examples) to the Github repository.

No news since last comment.

> What more detailed explanation are you missing for the log visualization? <- can this link or this information be found somewhere on the initial Github webpage

Added link and example image to the `README.md`.

>I think is important to include in those lines regarding the aspects such as:
a) the figures/plots it produces [..]
b) all the methods/models that can be learned [..]

I added a plot example to the paper and also explicitly named some algorithms as well as emphasize using scikit-learn algorithms (see [the latest article proof](https://github.com/openjournals/joss-papers/blob/joss.01132/joss.01132/10.21105.joss.01132.pdf)).
Furthermore I also explicitly wrote out and linked some of the algorithms used [in the documentation](https://pgijsbers.github.io/gama/user_guide/index.html#gama-search-space-configuration
).
However, I think including *all* the different plots produced and algorithms used is not a great idea for three reasons:
 1. There are many different plots and algorithms, taking up a lot of space while I think the current selection gives a good impression.
 2. Precise plots and algorithms are not what defines the package. You are free to extract more information from logs and/or add your own algorithms or remove used ones.
 3. Algorithms and visualization are susceptible to change. Other machine learning packages might be incorporated and/or better visualizations might be developed.

I find the adjustments I made are a reasonable balance between providing too little information and all information. Please let me know if you disagree.

> 4. There were .csv files generated that are not well described anywhere
> 4.1 There were folders empty folders created with the same name of the .csv files.
> 4.2- Can the user specify the folder/csv files names instead of automatically assigning the date of the experimental run?

They now have their section [in the documentation](https://pgijsbers.github.io/gama/user_guide/index.html#generated-files).
After reconsideration, the `.csv` is no longer be generated as `gama.log` supersedes it (and is thus not documented).
