### 2019-01-17 #1132 @jsgalan
Hi all @arfon @PGijsbers

I just revised and all the changes suggested were made in documents and pdf, respectively. I am happy to say that from my part all the requirements were fulfilled.

> However, I think including all the different plots produced and algorithms used is not a great idea for three reasons:
> There are many different plots and algorithms, taking up a lot of space while I think the current selection gives a good impression.
> Precise plots and algorithms are not what defines the package. You are free to extract more information from logs and/or add your own algorithms or remove used ones.
> Algorithms and visualization are susceptible to change. Other machine learning packages might be incorporated and/or better visualizations might be developed

Answering to this comment, after my initial thoughts and replies, I imagined that the package can be (and will be extended) to include different types of tests or can be put into a greater pipeline using various packages, so i concur with your reasoning.

Happy to revise and test this novel tool that will be beneficial for the ML community.

Best
