### 2019-01-18 7d218d9
[skip ci] Added documentation. Removed __main__ entry point.
