### 2019-01-18 18e098b
Terminal is now immutable.
