### 2019-01-18 11fcc50
Make Primitive immutable.
