### 2019-01-20 cbca87e
Changed syntax to be 3.5 compatible.
