### 2019-01-20 23b1548
Further refactoring.
