+++ tests/unit/test_automl_gp.py
@@ -1,6 +1,6 @@
 import unittest
 
-from gama.genetic_programming.components import Individual
+from gama.genetic_programming.components import Individual, Fitness
 from gama.genetic_programming.selection import eliminate_from_pareto
 from gama.configuration.testconfiguration import clf_config
 from gama import GamaClassifier
@@ -65,9 +65,9 @@ class AutomlGpTestCase(unittest.TestCase):
         self.assertEqual(len(list(self.individual_list[2].primitives)), 1)
 
     def test_eliminate_NSGA(self):
-        self.individual_list[0].fitness.values = (3, -2)
-        self.individual_list[1].fitness.values = (4, -2)
-        self.individual_list[2].fitness.values = (3, -1)
+        self.individual_list[0].fitness = Fitness((3, -2), 0, 0)
+        self.individual_list[1].fitness = Fitness((4, -2), 0, 0)
+        self.individual_list[2].fitness = Fitness((3, -1), 0, 0)
 
         eliminated = eliminate_from_pareto(pop=self.individual_list, n=1)
         self.assertListEqual(eliminated, [self.individual_list[0]],
