### 2019-01-20 f377a36
Further documentation, rename method.
