### 2019-01-20 09c428a
Refactor into smaller files.
