+++ gama/genetic_programming/components.py
@@ -1,318 +1,0 @@
-"""
-Defines the building blocks for individuals, in short these are the types:
-- Terminal. Definition of a specific value for a specific hyperparameter. Immutable.
-- Primitive. Definition of a specific algorithm; defined by Terminal input, output type and operation. Immutable.
-- PrimitiveNode. An instanciated Primitive with specific Terminals. Mutable for easy operations (e.g. mutation).
-- Individual. A sequence of PrimitiveNodes that together form a pipeline.
-    - Fitness. Stores information about the evaluation of the individual.
-
-"""
-from collections import defaultdict
-import random
-from typing import List, Callable, NamedTuple, Tuple
-import uuid
-
-import sklearn
-
-DATA_TERMINAL = 'data'
-
-# Specifies a specific value for a specific type or input, e.g. a value for a hyperparameter for an algorithm.
-Terminal = NamedTuple("Terminal",
-                      [("value", object),
-                       ("output", str),
-                       ("identifier", str)])
-
-
-def str_format_terminal_value(terminal) -> str:
-    if isinstance(terminal.value, str):
-        return "'{}'".format(terminal.value)
-    elif callable(terminal.value):
-        return "{}".format(terminal.value.__name__)
-    else:
-        return str(terminal.value)
-
-
-def terminal__str__(terminal) -> str:
-    """ e.g. "tol=0.5" """
-    return "{}={}".format(terminal.output, str_format_terminal_value(terminal))
-
-
-def terminal__repr__(terminal) -> str:
-    """ e.g. "FastICA.tol=0.5". Note that if the hyperparameter is shared across primitives, there is no prefix. """
-    return "{}={}".format(terminal.identifier, str_format_terminal_value(terminal))
-
-
-Terminal.__str__ = terminal__str__
-Terminal.__repr__ = terminal__repr__
-
-# Defines an operator which takes input and produces output, e.g. a preprocessing or classification algorithm.
-Primitive = NamedTuple("Primitive",
-                       [("input", Tuple[str]),
-                        ("output", str),
-                        ("identifier", Callable)])
-
-
-def primitive__str__(primitive) -> str:
-    """ e.g. "FastICA" """
-    return primitive.identifier.__name__
-
-
-Primitive.__str__ = primitive__str__
-Primitive.__repr__ = primitive__str__
-
-Fitness = NamedTuple("Fitness",
-                     [("values", Tuple),
-                      ("start_time", int),
-                      ("time", int)])
-
-
-class PrimitiveNode:
-    """ An instantiation for a Primitive with specific Terminals. """
-
-    def __init__(self, primitive: Primitive, data_node, terminals: List[Terminal]):
-        self._primitive = primitive
-        self._data_node = data_node
-        self._terminals = sorted(terminals, key=lambda t: str(t))
-
-    def __str__(self):
-        """ Recursively stringify all primitive nodes (primitive and hyperparameters).
-
-        Examples: - "GaussianNB(data)"
-                  - "BernoulliNB(data, alpha=1.0)"
-                  - "BernoulliNB(FastICA(data, tol=0.5), alpha=1.0)"
-        """
-        if self._terminals:
-            terminal_str = ", ".join([repr(terminal) for terminal in self._terminals])
-            return "{}({}, {})".format(self._primitive, str(self._data_node), terminal_str)
-        else:
-            return "{}({})".format(self._primitive, str(self._data_node))
-
-    def copy(self):
-        """ Make a shallow copy w.r.t. Primitive/Terminal (they are immutable), but deep w.r.t. PrimitiveNodes. """
-        data_node_copy = self._data_node if self._data_node == DATA_TERMINAL else self._data_node.copy()
[…]
