### 2019-01-20 #1132 @arokem
Great work everyone! Thanks for the review @jsgalan!

@PGijsbers : your paper is ready to be processed for acceptance.

Could you please create an archive of the current state of your software (e.g., by creating a tag/version of the software and uploading that to [zenodo](https://zenodon.org)).

Once you have a DOI for the software, please post it here, so we can add the archive to the paper.
