### 2019-01-21 #1132 @PGijsbers
Done. Zenodo directs me [here](https://zenodo.org/badge/latestdoi/115117161) but it doesn't show. I assume it will just take a few minutes?
