### 2019-01-21 #1132 @PGijsbers
Seems like it is all fixed @arokem
