### 2019-01-22 8b3fad4
Now record both wallclock time and process time of evaluation.
