+++ gama/genetic_programming/components/fitness.py
@@ -3,4 +3,5 @@ from typing import NamedTuple, Tuple
 Fitness = NamedTuple("Fitness",
                      [("values", Tuple),
                       ("start_time", int),
-                      ("time", int)])
+                      ("wallclock_time", int),
+                      ("process_time", int)])
