+++ gama/genetic_programming/compilers/scikitlearn.py
@@ -79,10 +79,10 @@ def object_is_valid_pipeline(o):
 
 def evaluate_individual(individual: Individual, operator_set: OperatorSet, evaluate_pipeline_length, *args, **kwargs):
     pipeline = operator_set.compile(individual)
-    (scores, start_datetime, evaluation_time) = evaluate_pipeline(pipeline, *args, **kwargs)
+    (scores, start_datetime, wallclock_time, process_time) = evaluate_pipeline(pipeline, *args, **kwargs)
     if evaluate_pipeline_length:
         scores = (*scores, -len(individual.primitives))
-    individual.fitness = Fitness(scores, start_datetime, evaluation_time)
+    individual.fitness = Fitness(scores, start_datetime, wallclock_time, process_time)
     return individual
 
 
@@ -124,7 +124,8 @@ def evaluate_pipeline(pl, X, y_train, y_score, timeout, metrics='accuracy', cv=5
             log.warning("File not found while saving predictions. This can happen in the multi-process case if the "
                         "cache gets deleted within `max_eval_time` of the end of the search process.", exc_info=True)
 
-    evaluation_time = time.process_time() - start
+    process_time = time.process_time() - start
+    wallclock_time = (datetime.now() - start_datetime).total_seconds()
 
     if c_mgr.state == c_mgr.INTERRUPTED:
         # A TimeoutException was raised, but not by the context manager.
@@ -139,5 +140,5 @@ def evaluate_pipeline(pl, X, y_train, y_score, timeout, metrics='accuracy', cv=5
         log_parseable_event(logger, TOKENS.EVALUATION_TIMEOUT, start_datetime, single_line_pipeline)
         logger.debug("Timeout after {}s: {}".format(timeout, pl))
 
-    fitness_values = (scores, start_datetime, evaluation_time)
+    fitness_values = (scores, start_datetime, wallclock_time, process_time)
     return fitness_values
