### 2019-01-23 35b521c
Squeeze predictions of shape (N,1) to (N,).
