### 2019-01-24 #1132 @PGijsbers
Actually not sure what is preferred here:
 - specifically this version: 10.5281/zenodo.2545472
 - link to latest: 10.5281/zenodo.2545471
