### 2019-01-24 8ae90e3
Check that array has 2nd dimension before accessing its shape.
