### 2019-01-24 94bcdd2
FunctionDispatcher is now a context manager
