+++ gama/d3m/metalearning.py
@@ -1,5 +1,9 @@
 from collections import Counter
 import numpy as np
+import pickle
+import os
+
+from ..genetic_programming.operations import create_seeded_individual
 
 
 def calculate_meta_features(X, y):
@@ -15,7 +19,7 @@ def calculate_meta_features(X, y):
 
     n_missing_values = np.isnan(X).sum()
     perc_missing_values = (n_missing_values / (n_instances * n_features)) * 100
-    n_instance_missing_values = np.isnan(X).sum(axis=1)
+    n_instance_missing_values = sum(np.isnan(X).any(axis=1))
     perc_instance_missing_values = (n_instance_missing_values / n_instances) * 100
 
     auto_correlation = 1 - np.mean([int(y2 != y1) for y1, y2 in zip(y, y[1:])])
@@ -37,20 +41,78 @@ def calculate_meta_features(X, y):
 
 
 def SVC_features(individual):
-    pass
+    svc_node = individual.main_node
+    kernel_encoder = dict(poly=0, rbf=1, sigmoid=2)
+    return [
+        [t.value for t in svc_node._terminals if t.output == 'C'][0],
+        [t.value for t in svc_node._terminals if t.output == 'coef0'][0],
+        [t.value for t in svc_node._terminals if t.output == 'degree'][0],
+        [t.value for t in svc_node._terminals if t.output == 'gamma'][0],
+        kernel_encoder[[t.value for t in svc_node._terminals if t.output == 'kernel'][0]],
+        [t.value for t in svc_node._terminals if t.output == 'shrinking'][0],
+        [t.value for t in svc_node._terminals if t.output == 'tol'][0]
+        ]
+
 
 def GradientBoosting_features(individual):
-    pass
+    gbc_node = individual.main_node
+    criterion_encoder = dict(friedman_mse=0, mae=1, mse=2)
+    return [
+        criterion_encoder[[t.value for t in gbc_node._terminals if t.output == 'criterion'][0]],
+        [t.value for t in gbc_node._terminals if t.output == 'learning_rate'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'max_depth'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'max_features'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'min_impurity_decrease'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'min_samples_leaf'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'min_samples_split'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'min_weight_fraction_leaf'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'n_estimators'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'subsample'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'tol'][0],
+        [t.value for t in gbc_node._terminals if t.output == 'validation_fraction'][0]
+        ]
+
 
 def RandomForestClassifier_features(individual):
-    pass
+    rfc_node = individual.main_node
+    criterion_encoder = dict(gini=1, entropy=0)
+    return [
+        [t.value for t in rfc_node._terminals if t.output == 'bootstrap'][0],
+        criterion_encoder[[t.value for t in rfc_node._terminals if t.output == 'criterion'][0]],
+        [t.value for t in rfc_node._terminals if t.output == 'max_features'][0],
+        [t.value for t in rfc_node._terminals if t.output == 'min_samples_leaf'][0],
+        [t.value for t in rfc_node._terminals if t.output == 'min_samples_split'][0]
+        ]
+
 
 def pick_best(metafeatures, configs, model, n):
-    pass
[…]
