+++ gama/genetic_programming/operations.py
@@ -27,3 +27,19 @@ def create_random_individual(primitive_set: dict, min_length: int = 1, max_lengt
         last_primitive_node = primitive_node
 
     return Individual(learner_node)
+
+
+def create_seeded_individual(primitive_set: dict, main: Primitive, min_length: int = 1, max_length: int = 3) -> Individual:
+    """ Create an individual with at least `min_length` Primitives and at most `max_length` Primitives that
+        has `main` as learner node. """
+    individual_length = random.randint(min_length, max_length)
+    terminals = random_terminals_for_primitive(primitive_set, main)
+    learner_node = PrimitiveNode(main, data_node=DATA_TERMINAL, terminals=terminals)
+    last_primitive_node = learner_node
+
+    for _ in range(individual_length - 1):
+        primitive_node = random_primitive_node(output_type=DATA_TERMINAL, primitive_set=primitive_set)
+        last_primitive_node._data_node = primitive_node
+        last_primitive_node = primitive_node
+
+    return Individual(learner_node)
