### 2019-01-25 5962ee3
Only warm start if learner is available in config.
