### 2019-01-25 518122e
Logging. Gotta find a better way to debug through the system.
