### 2019-01-26 c964dc7
Less complicated.
