### 2019-01-26 dc6f5f8
updated docstring.
