### 2019-01-26 07eb0bf
function for queueing all at once.
