### 2019-01-29 62c889f
gama log level should be set on loading the module, not when creating the instance.
