### 2019-01-29 f337d26
gama log level should be set on loading the module, not when creating the instance.
