+++ gama/gama.py
@@ -31,6 +31,10 @@ from gama.configuration.parser import pset_from_config
 from gama.genetic_programming.operator_set import OperatorSet
 from gama.genetic_programming.compilers.scikitlearn import compile_individual
 
+#  `gamalog` is for the entire gama module and submodules.
+gamalog = logging.getLogger('gama')
+gamalog.setLevel(logging.DEBUG)
+
 log = logging.getLogger(__name__)
 
 STR_NO_OPTIMAL_PIPELINE = """Gama did not yet establish an optimal pipeline.
@@ -104,9 +108,6 @@ class Gama(object):
                  keep_analysis_log=True,
                  cache_dir=None):
 
-        #  gamalog is for the entire gama module and submodules.
-        gamalog = logging.getLogger('gama')
-        gamalog.setLevel(logging.DEBUG)
         if verbosity >= logging.DEBUG:
             stdout_streamhandler = logging.StreamHandler(sys.stdout)
             stdout_streamhandler.setLevel(verbosity)
