### 2019-01-30 03b79aa
new d3m version requirements.
