+++ setup.py
@@ -3,9 +3,9 @@
 from setuptools import setup, find_packages
 
 requirements = [
-    'numpy==1.14.3',
+    'numpy==1.15.4',
     'scipy>=1.0.0',
-    'scikit-learn==0.19.1',
+    'scikit-learn==0.20.2',
     'stopit>=1.1.1',
     'liac-arff>=2.2.2',
     'category-encoders>=1.2.8'
@@ -13,7 +13,7 @@ requirements = [
 
 setup(
     name='gama',
-    version='19.01.0',
+    version='19.01.0.d3m',
     description='A package for automated machine learning based on scikit-learn.',
     long_description='',
     long_description_content_type='text/markdown',
